Extract Settings schema, field HTML, and sanitization into collaborators so field definitions and type rules can be unit-tested without booting WordPress.
