+++ abilities/class-ability-update-settings.php
@@ -103,13 +103,14 @@ class Ability_Update_Settings extends Ability {
 
 		// Build allowlist of {section}_{field} keys from registered settings,
 		// and a parallel list of which keys correspond to checkbox fields.
-		// We need the latter because Settings::sanitize_setting_by_field_type()
+		// We need the latter because Settings_Sanitizer::sanitize_setting_by_field_type()
 		// only accepts numeric values for checkboxes (is_numeric() rejects PHP
 		// booleans), so true/false from a JSON client would otherwise be
 		// silently coerced to '' instead of 0/1.
 		$valid_keys    = array();
 		$checkbox_keys = array();
-		foreach ( $this->plugin->settings->get_fields() as $section => $section_data ) {
+		$fields        = $this->plugin->settings->registry->get_fields();
+		foreach ( $fields as $section => $section_data ) {
 			if ( empty( $section_data['fields'] ) || ! is_array( $section_data['fields'] ) ) {
 				continue;
 			}
@@ -147,7 +148,7 @@ class Ability_Update_Settings extends Ability {
 		// Run only the incoming keys through Stream's sanitize pipeline so
 		// values are normalized to their declared field type, then merge over
 		// the existing options so unrelated keys are preserved.
-		$sanitized = $this->plugin->settings->sanitize_settings( $filtered );
+		$sanitized = $this->plugin->settings->sanitizer->sanitize_settings( $filtered, $fields );
 		$merged    = array_merge( $current, $sanitized );
 
 		// update_all_setting_values() persists to the correct store and
