+++ classes/class-admin-purge.php
@@ -495,14 +495,14 @@ class Admin_Purge {
 			return;
 		}
 
-		$defaults = $this->admin->plugin->settings->get_defaults();
+		$defaults = $this->admin->plugin->settings->registry->get_defaults();
 		if ( $this->admin->plugin->is_multisite_network_activated() ) {
 			$options = wp_parse_args( (array) get_site_option( 'wp_stream_network', array() ), $defaults );
 		} else {
 			$options = wp_parse_args( (array) get_option( 'wp_stream', array() ), $defaults );
 		}
 
-		// TTL fallback. Settings::get_defaults() runs every settings field
+		// TTL fallback. Settings_Registry::get_defaults() runs every settings field
 		// through the `wp_stream_settings_option_fields` filter, which
 		// Network::get_network_admin_fields() uses to strip the `records_ttl`
 		// field from the per-site option's defaults set. When this callback runs
